feat(md-extras): move mermaid/image-enlarge/admonitions-preset/toc to zfb-md-extras (#570)
Wave 3 sub-issue: migrate 4 opt-in framework features from zfb-content to
zfb-md-extras and make Pipeline::with_defaults_and_features conditional on
the corresponding MarkdownFeaturesConfig flags.

Key changes:
- Move DirectiveDef/AttrSchema/AttrType/etc. to zfb-md-ast (prerequisite for
  admonitions_preset to live in zfb-md-extras without a dep cycle)
- Move extract_text helper to zfb-md-ast for the same reason
- Add zfb-md-extras modules: mermaid, image_enlarge, admonitions_preset,
  heading_marker_toc (each with full implementation + unit tests)
- Replace old plugin files in zfb-content with re-export shims pointing to
  zfb-md-extras (backwards-compatible import paths preserved)
- Rewrite Pipeline::with_defaults_and_features to build the framework chain
  without the 4 opt-in plugins; register_features conditionally adds them
- Flip the byte-identical test to assert that empty features does NOT wire
  the opt-in plugins
- Add 4 doc pages under docs/src/content/docs/markdown-features/ with tier: Opt-in
