+++ docs/src/content/docs/markdown-features/image-enlarge.mdx
@@ -1,0 +1,43 @@
+---
+title: Image enlarge
+description: Wrap block-level images in an enlargeable figure with a zoom button.
+tier: Opt-in
+---
+
+The `imageEnlarge` feature wraps block-level paragraph images in a `<figure class="zd-enlargeable">` element with an SVG zoom button, enabling client-side lightbox behaviour.
+
+## Enable
+
+```ts
+// zfb.config.ts
+export default defineConfig({
+  markdown: {
+    features: {
+      imageEnlarge: true,
+    },
+  },
+});
+```
+
+## Behaviour
+
+**Selector:** any `<p>` whose only non-whitespace child is an `<img>` element. Images inside `<a>`, `<picture>`, or paragraphs with surrounding text are left untouched.
+
+**Output shape:**
+
+```html
+<figure class="zd-enlargeable">
+  <img src="photo.png" alt="..." />
+  <button type="button" class="zd-enlarge-btn" aria-label="Enlarge image" hidden>
+    <!-- 4-polygon SVG icon -->
+  </button>
+</figure>
+```
+
+**Opt-out:** add `title="no-enlarge"` to the `<img>` to skip wrapping. The sentinel attribute is stripped from the rendered `<img>` so it never reaches the DOM:
+
+```md
+![alt text](photo.png "no-enlarge")
+```
+
+**Idempotency:** a paragraph whose only child is an already-wrapped `<figure class="zd-enlargeable">` is left alone.
